+++ crates/base/core/src/run/time/frame/mod.rs
@@ -1,18 +1,20 @@
-// devela_base_core::run::frame
+// devela_base_core::run::time::frame
 //
 #![doc = crate::_tags!(wip)]
-#![doc = crate::_DOC_RUN_FRAME!()] // public
-#![doc = crate::_doc!(modules: crate::run; frame)]
+#![doc = crate::_DOC_RUN_TIME_FRAME!()] // public
+#![doc = crate::_doc!(modules: crate::run::time; frame)]
 #![doc = crate::_doc!(flat:"run")]
 #![doc = crate::_doc!(hr)]
 //
 
-// mod pacer; // FramePacer WIP
+mod pacer; // FramePacer
+mod step; // RunStep
 
 crate::structural_mods! { // _mods
     _mods {
-        // pub use super::{
-        //     pacer::*,
-        // };
+        pub use super::{
+            pacer::*,
+            step::*,
+        };
     }
 }