+++ src/run/cycle/mod.rs
@@ -1,0 +1,14 @@
+// devela::run::cycle
+//
+#![doc = crate::_DOC_RUN_CYCLE!()] // public
+#![doc = crate::_doc!(modules: crate::run; cycle)]
+#![doc = crate::_doc!(flat:"run")]
+#![doc = crate::_doc!(hr)]
+
+crate::structural_mods! { // _reexports
+    _reexports {
+        pub use devela_base_core::run::cycle::{
+            RunControl, RunCycle, RunPhase
+        };
+    }
+}